+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - basic autodownloading of arXiv PDF files (#25,!35):
     - the default download location can be configured via `config.utils.file_downloader.default_location`
     - on a per addition basis, this default can be overwritten via the `--path` keyword option of the `AddCommand`
+- automatic journal abbreviations (#62,!36):
+    - users can configure a list of journal abbreviations via `config.utils.journal_abbreviations`
+    - if present, coBib will store the journal in its elongated form
+    - a user can then automatically convert to abbreviated forms during exporting (see the new `--abbreviate` and `--dotless` arguments)
 
 ## [3.1.1] - 2021-05-25